My fixes (#3)
* Updated database and added new code

Updated the item database to that of CrossCode v1.0.0-9.
Also added a section to edit the list of favorite consumables. (My save was bugged for some reason and was auto-favoriting things that shouldn't be put there)

* Update database, added batch script.

Updated the item database to that of 1.0.0-10.
Also added a Windows Batch script to update the "itemj.js" if the "item-database.json" has been copied from the game assets into the editor's folder.

* Limit Favorite Consumables

Limited the list of Favorite Consumables to 12. Now if the save happens to have more than that, the extras are simply discarded.

* Fixed bug with Favorites

Fixed a bug where previous Favorites were still there after deleting some Favorites.

* Resolved conflict in PR

* Slight code cleanup.

Removed a leftover line of code from before the limiting of the favourites list.

Actually surprised it didn't break anything with regards to the list of Favourite Consumables...

* Delete makeitemj.cmd

* Delete itemj.js.stub

* Delete itemj.js